Added some suggested scores.
Moved the RBL query method domains into variables so they could be wrapped in if statements

Explained to use either local files or rbl to query sendgrid.

This file should be usable as is and not break if the user hasn't downloaded the files or set the rbl lookup domains.